MLOps/mlops-code/data_pipeline/dev_requirements.txt
apache-airflow==2.3.4
apache-airflow-providers-docker==3.1.0
feast[redis]==0.24.0
pandas==1.4.4
pendulum==2.1.2
pyspark==3.0.1
firebase-admin==6.5.0